+++ src/nexgen/tools/ED_tools.py
@@ -7,6 +7,8 @@ from pathlib import Path
 from typing import Any, Dict, List, Tuple
 
 import h5py
+import hdf5plugin  # noqa: F401
+import numpy as np
 from numpy.typing import ArrayLike
 
 try:
@@ -153,7 +155,7 @@ def extract_from_SINGLA_master(master: Path | str) -> Dict[str, Any]:
     from a Singla master file.
 
     Args:
-        master (Path | str): Path to master file.
+        master (Path | str): Path to Singla master file.
 
     Returns:
         Dict[str, Any]: Dictionary of information relative to the detector.
@@ -187,3 +189,78 @@ def extract_from_SINGLA_master(master: Path | str) -> Dict[str, Any]:
         D["photon_energy"] = singla.get_photon_energy()
 
     return D
+
+
+def centroid_max(image: ArrayLike) -> Tuple[float, float]:
+    """
+    Find the centre of gravity of the maximum pixels.
+
+    Args:
+        image (ArrayLike): Pixel image.
+
+    Returns:
+        Tuple[float, float]: Centroid (x,y) position.
+    """
+
+    y, x = np.where(image == np.amax(image))
+    return np.mean(x), np.mean(y)
+
+
+def find_beam_centre(
+    master: Path | str, data: Path | str, data_entry_key: str = "/entry/data/data"
+) -> Tuple[float, float]:
+    """
+    Calculate the beam center position for Electron Diffraction data collected on Singla detector.
+
+    Args:
+        master (Path | str): Path to Singla master file.
+        data (Path | str): Path to data file.
+        data_entry_key (str, optional): Key for the location of the images inside the Singla data file. Defaults to "/entry/data/data".
+
+    Returns:
+        fast, slow (Tuple[float, float]): Beam center position (fast, slow) on the detector. \
+            None if the pixel_mask can't be found.
+    """
+
+    with h5py.File(master, "r") as fh:
+        singla = SinglaMaster(fh)
+        pixel_mask = singla.get_mask()[1]
+
+    # If the module gap is unmasked we would get bad results
+    if pixel_mask is None:
+        return None
+
+    # Set the ROI to be +/- 100 pixels around the image centre
+    yc, xc = (e // 2 for e in pixel_mask.shape)
+    x0 = xc - 100
+    x1 = xc + 100
+    y0 = yc - 100
+    y1 = yc + 100
+
+    # Bool selection for masked pixels in the ROI
+    pixel_mask = pixel_mask[y0:y1, x0:x1] == 1
+
+    images = []
+    with h5py.File(data, "r") as fh:
+        data = fh[data_entry_key]
+        num_images = data.shape[0]
+        for i in range(0, num_images, num_images // min(num_images, 10)):
+            image = data[i, y0:y1, x0:x1]
+            image[pixel_mask] = 0
+            images.append(image)
+
+    beam_centres = [centroid_max(im) for im in images]
+    x, y = zip(*beam_centres)
+
+    # For robustness against blank images, remove any value more than 5 px
+    # from the median
+    med_x = np.median(x)
+    med_y = np.median(y)
+    x = [e for e in x if abs(e - med_x) < 5]
+    y = [e for e in y if abs(e - med_y) < 5]
+
+    # Correct for offset of the ROI and shift to centre pixel
+    fast = xc - 100 + np.mean(x) + 0.5
+    slow = yc - 100 + np.mean(y) + 0.5
+
+    return fast, slow